requirements.txt
numpy>=2.0.2
openai>=1.59.6
openai-whisper>=20240930
sounddevice>=0.5.1
soundfile>=0.13.0
streamlit>=1.41.1
psycopg2-binary>=2.9.10
nltk>=3.9.1
reportlab>=4.2.5
pdfgen>=1.0.5
google-cloud-storage>=2.19.0
ipykernel>=6.29.5
